pyproject.toml
[tool.poetry]
name = "nlavida"
version = "0.1.0"
description = "Natural Language-Assisted Visualization and Interactive Data Analysis"
authors = ["obaidur-rahaman <ramieor@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [
    { include = "backend" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
anyio = "4.3.0"
argon2-cffi = "23.1.0"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
debugpy = "1.8.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
entrypoints = "0.4"
flask = "3.0.3"
idna = "3.7"
importlib-metadata = "7.1.0"
ipykernel = "6.29.3"
ipython = ">=7.0.0,<8.25.0"
ipywidgets = "8.1.2"
jedi = "0.19.1"
jinja2 = "3.1.3"
joblib = "1.4.2"
jsonschema = "4.22.0"
jupyter = "1.0.0"
jupyter_client = "8.6.1"
jupyter_console = "6.6.3"
jupyter_core = "5.7.2"
jupyterlab = "4.1.8"
kiwisolver = "1.4.5"
langchain_experimental = "0.0.58"
langchain = "0.1.19"
langchain-community = "0.0.38"
langchain-core = "0.1.52"
langchain-openai = "0.1.4"
langchainhub = "0.1.15"
langgraph = "0.0.44"
markupsafe = "2.1.5"
matplotlib = "3.8.4"
mistune = "3.0.2"
nbclient = "0.10.0"
nbconvert = "7.16.4"
nbformat = "5.10.4"
notebook = "7.1.3"
numpy = "1.26.4"
pandas = "1.5.3"
pillow = "10.3.0"
prometheus_client = "0.20.0"
pygments = "2.17.2"
pytest = ">=8.2.1,<9.0.0"
pytest-flask = "^1.3.0"
pytest-mock = "^3.14.0"
python-dotenv = "^0.21.0"
pyyaml = "6.0.1"
requests = "2.31.0"
scikit-learn = "1.4.2"
scipy = "1.13.0"
seaborn = "0.13.2"
setuptools = "69.5.1"
six = "1.16.0"
sqlalchemy = "2.0.29"
statsmodels = "0.14.1"
tornado = "6.4"
tqdm = "4.66.2"
traitlets = "5.14.3"
zipp = "3.17.0"
flask-cors = "^4.0.1"
more-itertools = "^10.3.0"
groq = "^0.9.0"
langchain-groq = "^0.1.5"

[tool.poetry.group.dev.dependencies]
astor = "0.8.1"
curl-cffi = "0.6.3"
duckdb = "0.10.2"
duckduckgo-search = "5.3.0"
faker = "19.13.0"
regex = "2024.4.16"
tiktoken = "0.6.0"
types-requests = "2.31.0.20240406"
tox = { version = "^4.15.0", python = ">=3.8" }
pytest-mock = "^3.14.0"
